backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
typing-extensions
openai
loguru
pytest
pytest-asyncio
httpx
python-dotenv